+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/entity/developer/DeveloperStatusType.java
@@ -37,4 +37,8 @@ public enum DeveloperStatusType {
         }
         return result;
     }
+
+    public String getName() {
+        return name;
+    }
 }