+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/validation/surveillance/reviewer/Reviewer.java
@@ -1,0 +1,8 @@
+package gov.healthit.chpl.validation.surveillance.reviewer;
+
+import gov.healthit.chpl.domain.surveillance.Surveillance;
+
+public interface Reviewer {
+
+    void review(Surveillance surveillance);
+}