+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/manager/impl/SurveillanceManagerTest.java
@@ -5,10 +5,14 @@ import java.util.Iterator;
 
 import javax.persistence.EntityNotFoundException;
 
+import org.ff4j.FF4j;
+import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.Mockito;
+import org.mockito.MockitoAnnotations;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -24,6 +28,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.github.springtestdbunit.DbUnitTestExecutionListener;
 import com.github.springtestdbunit.annotation.DatabaseSetup;
 
+import gov.healthit.chpl.FeatureList;
 import gov.healthit.chpl.auth.permission.GrantedPermission;
 import gov.healthit.chpl.auth.user.JWTAuthenticatedUser;
 import gov.healthit.chpl.caching.UnitTestRules;
@@ -67,6 +72,9 @@ public class SurveillanceManagerTest extends TestCase {
     @Autowired
     public UnitTestRules cacheInvalidationRule;
 
+    @Autowired
+    private FF4j ff4j;
+
     private static JWTAuthenticatedUser adminUser;
     private static JWTAuthenticatedUser acbUser;
     private static JWTAuthenticatedUser acbUser2;
@@ -103,6 +111,12 @@ public class SurveillanceManagerTest extends TestCase {
         atlUser.getPermissions().add(new GrantedPermission("ROLE_ATL"));
     }
 
+    @Before
+    public void setup() {
+        MockitoAnnotations.initMocks(this);
+        Mockito.doReturn(true).when(ff4j).check(FeatureList.EFFECTIVE_RULE_DATE_PLUS_ONE_WEEK);
+    }
+
     @Test
     @Transactional
     @Rollback(true)
@@ -132,7 +146,7 @@ public class SurveillanceManagerTest extends TestCase {
 
         boolean failed = false;
         try {
-            Long insertedId = survManager.createSurveillance(-1L, surv);
+            Long insertedId = survManager.createSurveillance(surv);
             assertNull(insertedId);
         } catch (AccessDeniedException ex) {
             System.out.println(ex.getClass() + ": " + ex.getMessage());
@@ -174,7 +188,7 @@ public class SurveillanceManagerTest extends TestCase {
 
         boolean failed = false;
         try {
-            Long insertedId = survManager.createSurveillance(-1L, surv);
+            Long insertedId = survManager.createSurveillance(surv);
             assertNull(insertedId);
         } catch (AccessDeniedException ex) {
             System.out.println(ex.getClass() + ": " + ex.getMessage());
@@ -216,7 +230,7 @@ public class SurveillanceManagerTest extends TestCase {
 
         Long insertedId;
         try {
-            insertedId = survManager.createSurveillance(-1L, surv);
+            insertedId = survManager.createSurveillance(surv);
             assertNotNull(insertedId);
             Surveillance got = survManager.getById(insertedId);
             assertNotNull(got);
@@ -284,7 +298,7 @@ public class SurveillanceManagerTest extends TestCase {
 
         Long insertedId;
         try {
-            insertedId = survManager.createSurveillance(-1L, surv);
+            insertedId = survManager.createSurveillance(surv);
             assertNotNull(insertedId);
             Surveillance got = survManager.getById(insertedId);
             assertNotNull(got);
@@ -340,10 +354,10 @@ public class SurveillanceManagerTest extends TestCase {
 
         Long insertedId;
         try {
-            insertedId = survManager.createSurveillance(-1L, surv);
+            insertedId = survManager.createSurveillance(surv);
             Surveillance insertedSurv = survManager.getById(insertedId);
             assertNotNull(insertedId);
-            survManager.deleteSurveillance(-1L, insertedSurv);
+            survManager.deleteSurveillance(insertedSurv);
             boolean failed = false;
             try {
                 survManager.getById(insertedId);
@@ -371,6 +385,7 @@ public class SurveillanceManagerTest extends TestCase {
         cp.setEdition(cp.getEdition());
         surv.setCertifiedProduct(cp);
         surv.setStartDate(new Date());
+        surv.setEndDate(new Date(surv.getStartDate().getTime() + 1000));
         surv.setRandomizedSitesUsed(10);
         SurveillanceType type = survDao.findSurveillanceType("Randomized");
         surv.setType(type);
@@ -384,7 +399,7 @@ public class SurveillanceManagerTest extends TestCase {
         surv.getRequirements().add(req);
         surv.setAuthority(Authority.ROLE_ACB);
 
-        Long insertedId = survManager.createSurveillance(-1L, surv);
+        Long insertedId = survManager.createSurveillance(surv);
         assertNotNull(insertedId);
 
         Surveillance got = survManager.getById(insertedId);
@@ -398,7 +413,7 @@ public class SurveillanceManagerTest extends TestCase {
         resType = survDao.findSurveillanceResultType("No Non-Conformity");
         req2.setResult(resType);
         got.getRequirements().add(req2);
-        survManager.updateSurveillance(-1L, got);
+        survManager.updateSurveillance(got);
 
         got = survManager.getById(insertedId);
         assertNotNull(got);
@@ -421,6 +436,7 @@ public class SurveillanceManagerTest extends TestCase {
         cp.setEdition(cp.getEdition());
         surv.setCertifiedProduct(cp);
         surv.setStartDate(new Date());
+        surv.setEndDate(new Date(surv.getStartDate().getTime() + 1000));
         surv.setRandomizedSitesUsed(10);
         SurveillanceType type = survDao.findSurveillanceType("Randomized");
         surv.setType(type);
@@ -441,7 +457,7 @@ public class SurveillanceManagerTest extends TestCase {
         req2.setResult(resType);
         surv.getRequirements().add(req2);
 
-        Long insertedId = survManager.createSurveillance(-1L, surv);
+        Long insertedId = survManager.createSurveillance(surv);
         assertNotNull(insertedId);
 
         Surveillance got = survManager.getById(insertedId);
@@ -449,7 +465,7 @@ public class SurveillanceManagerTest extends TestCase {
         assertEquals(2, got.getRequirements().size());
 
         got.getRequirements().remove(got.getRequirements().iterator().next());
-        survManager.updateSurveillance(-1L, got);
+        survManager.updateSurveillance(got);
         got = survManager.getById(insertedId);
         assertNotNull(got);
         assertEquals(1, got.getRequirements().size());
@@ -471,6 +487,7 @@ public class SurveillanceManagerTest extends TestCase {
         cp.setEdition(cp.getEdition());
         surv.setCertifiedProduct(cp);
         surv.setStartDate(new Date());
+        surv.setEndDate(new Date(surv.getStartDate().getTime() + 1000));
         surv.setRandomizedSitesUsed(10);
         SurveillanceType type = survDao.findSurveillanceType("Randomized");
         surv.setType(type);
@@ -484,7 +501,7 @@ public class SurveillanceManagerTest extends TestCase {
         surv.getRequirements().add(req);
         surv.setAuthority(Authority.ROLE_ACB);
 
-        Long insertedId = survManager.createSurveillance(-1L, surv);
+        Long insertedId = survManager.createSurveillance(surv);
         assertNotNull(insertedId);
 
         Surveillance got = survManager.getById(insertedId);
@@ -504,7 +521,7 @@ public class SurveillanceManagerTest extends TestCase {
         SurveillanceNonconformityStatus ncStatus = survDao.findSurveillanceNonconformityStatusType("Open");
         nc.setStatus(ncStatus);
         gotReq.getNonconformities().add(nc);
-        survManager.updateSurveillance(-1L, got);
+        survManager.updateSurveillance(got);
 
         got = survManager.getById(insertedId);
         assertNotNull(got);
@@ -557,7 +574,7 @@ public class SurveillanceManagerTest extends TestCase {
         req.getNonconformities().add(nc);
         surv.setAuthority(Authority.ROLE_ACB);
 
-        Long insertedId = survManager.createSurveillance(-1L, surv);
+        Long insertedId = survManager.createSurveillance(surv);
         assertNotNull(insertedId);
 
         Surveillance got = survManager.getById(insertedId);
@@ -567,8 +584,8 @@ public class SurveillanceManagerTest extends TestCase {
         resType = survDao.findSurveillanceResultType("No Non-Conformity");
         gotReq.setResult(resType);
         gotReq.getNonconformities().clear();
-
-        survManager.updateSurveillance(-1L, got);
+        got.setEndDate(new Date(got.getStartDate().getTime() + 1000));
+        survManager.updateSurveillance(got);
 
         got = survManager.getById(insertedId);
         assertNotNull(got);
@@ -594,6 +611,7 @@ public class SurveillanceManagerTest extends TestCase {
         cp.setEdition(cp.getEdition());
         surv.setCertifiedProduct(cp);
         surv.setStartDate(new Date());
+        surv.setEndDate(new Date(surv.getStartDate().getTime() + 1000));
         surv.setRandomizedSitesUsed(10);
         SurveillanceType type = survDao.findSurveillanceType("Randomized");
         surv.setType(type);
@@ -607,14 +625,13 @@ public class SurveillanceManagerTest extends TestCase {
         surv.getRequirements().add(req);
         surv.setAuthority(Authority.ROLE_ACB);
 
-        Long insertedId = survManager.createSurveillance(-1L, surv);
+        Long insertedId = survManager.createSurveillance(surv);
         assertNotNull(insertedId);
 
         Surveillance got = survManager.getById(insertedId);
         assertNotNull(got);
-        assertNull(got.getEndDate());
         got.setEndDate(new Date());
-        survManager.updateSurveillance(-1L, got);
+        survManager.updateSurveillance(got);
 
         got = survManager.getById(insertedId);
         assertNotNull(got);
@@ -625,7 +642,7 @@ public class SurveillanceManagerTest extends TestCase {
 
     /**
      * OCD-1810.
-     * 
+     *
      * @throws EntityRetrievalException
      *             if entity can't be retrieved
      */
@@ -676,7 +693,7 @@ public class SurveillanceManagerTest extends TestCase {
 
         Long insertedId;
         try {
-            insertedId = survManager.createSurveillance(-1L, surv);
+            insertedId = survManager.createSurveillance(surv);
             assertNotNull(insertedId);
             Surveillance got = survManager.getById(insertedId);
             assertNotNull(got.getErrorMessages());